+++ .travis.yml
@@ -47,6 +47,7 @@ script:
   - sonar-scanner
   - cd $TRAVIS_BUILD_DIR/libexec/python && python2 -m unittest tests.test_json
   - cd $TRAVIS_BUILD_DIR/libexec/python && python2 -m unittest tests.test_base
+  - cd $TRAVIS_BUILD_DIR/libexec/python && python2 -m unittest tests.test_core
   - cd $TRAVIS_BUILD_DIR/libexec/python && python2 -m unittest tests.test_docker_import
   - cd $TRAVIS_BUILD_DIR/libexec/python && python2 -m unittest tests.test_docker_import_user
   - cd $TRAVIS_BUILD_DIR/libexec/python && python2 -m unittest tests.test_docker_api
@@ -58,6 +59,7 @@ script:
   - cd $TRAVIS_BUILD_DIR/libexec/python && python2 -m unittest tests.test_disable_cache
   - cd $TRAVIS_BUILD_DIR/libexec/python && python3 -m unittest tests.test_json
   - cd $TRAVIS_BUILD_DIR/libexec/python && python3 -m unittest tests.test_base
+  - cd $TRAVIS_BUILD_DIR/libexec/python && python3 -m unittest tests.test_core
   - cd $TRAVIS_BUILD_DIR/libexec/python && python3 -m unittest tests.test_docker_import
   - cd $TRAVIS_BUILD_DIR/libexec/python && python3 -m unittest tests.test_docker_import_user
   - cd $TRAVIS_BUILD_DIR/libexec/python && python3 -m unittest tests.test_docker_api
