+++ libexec/bootstrap-scripts/environment/shell
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-for script in /.singularity/env/*.sh; do
+for script in /singularity.d/env/*.sh; do
     if [ -f "$script" ]; then
         . $script
     fi
