+++ libexec/python/tests/test_shub_import.py
@@ -49,9 +49,8 @@ class TestImport(TestCase):
         # Variables are obtained from environment
         os.environ["SINGULARITY_CONTAINER"] = "shub://vsoch/singularity-images" 
         os.environ["SINGULARITY_ROOTFS"] = self.tmpdir
-        os.mkdir('%s/.singularity' %(self.tmpdir))
-        os.mkdir('%s/.singularity/labels' %(self.tmpdir))
-
+        os.mkdir('%s/singularity.d' %(self.tmpdir))
+        
         print("\n---START----------------------------------------")
 
     def tearDown(self):
