+++ src/action-lib/shell.c
@@ -35,10 +35,10 @@ void action_shell(int argc, char **argv) {
 
     singularity_message(INFO, "Singularity: Invoking an interactive shell within container...\n\n");
 
-    if ( is_exec("/.singularity/actions/shell") == 0 ) {
-        singularity_message(DEBUG, "Exec'ing /.singularity/actions/shell\n");
-        if ( execv("/.singularity/actions/shell", argv) < 0 ) { // Flawfinder: ignore
-            singularity_message(ERROR, "Failed to execv() /.singularity/actions/shell, continuing to /bin/sh: %s\n", strerror(errno));
+    if ( is_exec("/singularity.d/actions/shell") == 0 ) {
+        singularity_message(DEBUG, "Exec'ing /singularity.d/actions/shell\n");
+        if ( execv("/singularity.d/actions/shell", argv) < 0 ) { // Flawfinder: ignore
+            singularity_message(ERROR, "Failed to execv() /singularity.d/actions/shell, continuing to /bin/sh: %s\n", strerror(errno));
         }
     } else if ( is_exec("/.shell") == 0 ) {
         singularity_message(DEBUG, "Exec'ing /.shell\n");
